feat: WIP - add filters and pagination to stores table
