+++ manifest.json
@@ -1,7 +1,7 @@
 {
   "name": "glovo-integration",
   "vendor": "vtex",
-  "version": "3.0.0",
+  "version": "2.0.13",
   "title": "Glovo Integration",
   "description": "Glovo Integration",
   "dependencies": {
