+++ react/graphql/getAppSettings.gql
@@ -6,6 +6,7 @@ query getGlovoIntegrationSettings {
       id
       storeName
       sellerId
+      sellerName
       affiliateId
       salesChannel
       postalCode
